requirements.txt
flask
gunicorn
sqlalchemy
psycopg2-binary
python-dotenv
werkzeug
pillow
requests
